Break out InteractiveSessionInputPart class
